feat: Make Plots into an Extension
