+++ Project.toml
@@ -11,14 +11,13 @@ DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
 FunctionWrappers = "069b7b12-0de2-55c6-9aab-29f3d0a68a2e"
 HomotopyContinuation = "f213a82b-91d6-5c5d-acf7-10f1c761b327"
 JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
-Latexify = "23fbe1c1-3f47-55db-b15f-69d7ec21a316"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 OrderedCollections = "bac558e1-5e72-5ebc-8fee-abe8a469f55d"
-Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 PrecompileTools = "aea7be01-6a6a-4083-8856-8a6e6704d82a"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+SciMLBase = "0bca4576-84f4-4d90-8ffe-ffa030f20462"
 SymbolicUtils = "d1185830-fcd6-423d-90d6-eec64667417b"
 Symbolics = "0c5d862f-8b57-4792-8d23-62f2024744c7"
 
@@ -26,10 +25,12 @@ Symbolics = "0c5d862f-8b57-4792-8d23-62f2024744c7"
 ModelingToolkit = "961ee093-0014-501f-94e3-6117800e7a78"
 NaNMath = "77ba4419-2d1f-58cd-9bb1-8ffee604a2e3"
 OrdinaryDiffEqTsit5 = "b1df2697-797e-41e3-8120-5422d3b24e4a"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 SteadyStateDiffEq = "9672c7b4-1e72-59bd-8a11-6ac3964bc41f"
 
 [extensions]
 ModelingToolkitExt = "ModelingToolkit"
+PlotsExt = "Plots"
 SteadyStateDiffEqExt = "SteadyStateDiffEq"
 TimeEvolution = "OrdinaryDiffEqTsit5"
 
@@ -41,15 +42,15 @@ DocStringExtensions = "0.9.3"
 FunctionWrappers = "1.1.3"
 HomotopyContinuation = "2.12"
 JLD2 = "0.4.48, 0.5"
-Latexify = "0.16"
 ModelingToolkit = "9.60"
-OrdinaryDiffEqTsit5 = "1.1"
+NaNMath = "1.1"
 OrderedCollections = "1.6"
+OrdinaryDiffEqTsit5 = "1.1"
 Plots = "1.40.9"
 PrecompileTools = "1.2"
 ProgressMeter = "1.7.2"
+SciMLBase = "2.72.0"
 SteadyStateDiffEq = "2.4.1"
 SymbolicUtils = "3.11"
 Symbolics = "6.23"
 julia = "1.10.0"
-NaNMath = "1.1"
