+++ docs/src/examples/state_dependent_perturbation.md
@@ -228,7 +228,7 @@ harmonic_tmp.equations = HB.Symbolics.substitute(
     HB.rearrange_standard(harmonic_normal).equations[1:2], Dict(u2 => ua, v2 => va)
 )
 harmonic_tmp.parameters = push!(harmonic_tmp.parameters, ua, va)
-prob = HarmonicBalance.Problem(harmonic_tmp)
+prob = HarmonicBalance.Problem(harmonic_tmp, varied, fixed)
 ````
 
 We will sweep over the $\omega-\lambda$ plane and substitute the non-zero amplitude solution of the antisymmetric mode into the coupled equations of thesymmetric mode.
@@ -267,8 +267,9 @@ function solve_perturbed_system(prob, input)
     rounded_solutions = HB.HomotopyContinuation.solutions.(result_full)
     solutions = HB.pad_solutions(rounded_solutions)
 
+    jacobian = HarmonicBalance.get_Jacobian(harmonic_tmp)
     J_variables = cat(prob.variables, collect(keys(varied)), [ua, va]; dims=1)
-    compiled_J = HB.compile_matrix(prob.jacobian, J_variables; rules=fixed)
+    compiled_J = HB.compile_matrix(jacobian, J_variables; rules=fixed)
     compiled_J = HB.JacobianFunction(HB.solution_type(solutions))(compiled_J)
     result = HB.Result(
         solutions,
