+++ examples/parametron.jl
@@ -9,7 +9,7 @@
 
 # To implement this system in Harmonic Balance, we first import the library
 
-using HarmonicBalance
+using HarmonicBalance, Plots
 
 # Subsequently, we type define parameters in the problem and the oscillating amplitude function $x(t)$ using the `variables` macro from `Symbolics.jl`
