+++ src/Result.jl
@@ -86,3 +86,5 @@ Base.size(r::Result) = size(r.solutions)
 
 branch_count(r::Result) = length(r.solutions[1])
 get_branch(r::Result, idx) = getindex.(r.solutions, idx)
+
+dim(res::Result) = length(size(res.solutions)) # give solution dimensionality
