+++ src/modules/LinearResponse/LinearResponse.jl
@@ -4,12 +4,8 @@ using Printf: Printf, @printf
 using DocStringExtensions
 using ProgressMeter: ProgressMeter, Progress, next!
 
-using Plots: heatmap, theme_palette, scatter, RGB, cgrad
-using Latexify: Latexify, latexify, @L_str
-using Latexify.LaTeXStrings: LaTeXStrings
-
 using Symbolics: Symbolics, Num, unwrap
-using LinearAlgebra: norm, eigvals, eigen, eigvecs
+using LinearAlgebra: norm, eigen
 
 using HarmonicBalance
 using HarmonicBalance:
@@ -26,10 +22,6 @@ using HarmonicBalance:
     d,
     substitute_all,
     _free_symbols,
-    _get_mask,
-    _set_Plots_default,
-    dim,
-    _get_mask,
     harmonic_ansatz,
     slow_flow,
     fourier_transform,
@@ -44,6 +36,5 @@ include("response.jl")
 include("plotting.jl")
 
 export get_Jacobian, show
-export plot_linear_response, plot_rotframe_jacobian_response, plot_eigenvalues
 
 end
